Prime_Leads/requirements.txt
langgraph>=0.0.40
flasgger>=0.9.7.1
openai>=1.0.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
flask>=2.3.0
requests>=2.31.0
pydantic>=2.0
numpy
google-generativeai>=0.3.2
